Using a Pipeline Approach
This notebook reuses much of the code if the original "Building the Model" notebook. However, here I used a Pipeline and the OneHotEncoder from sklearn.preprocessing to train and fit my model. The result is exponentially better than the original approach. I expect it to be due to using the OneHotEncoder over using my self-defined binnings.